remove support foot and remove default for support side
