+++ crates/types/src/initial_look_around.rs
@@ -1,6 +1,6 @@
 use serde::{Deserialize, Serialize};
 
-use crate::support_foot::Side;
+use crate::support_side::Side;
 
 #[derive(Debug, Clone, Copy, Serialize, Deserialize, PartialEq, ros_z::Message)]
 pub enum BallSearchLookAround {
