+++ crates/types/src/initial_pose.rs
@@ -1,6 +1,6 @@
 use serde::{Deserialize, Serialize};
 
-use crate::support_foot::Side;
+use crate::support_side::Side;
 
 #[derive(Clone, Debug, Default, Deserialize, Serialize, ros_z::Message)]
 pub struct InitialPose {
